Build 7

+++ Assets/_Assets/u1w202012/SilCil/Scripts/NewPieceManager/NewCombinationScoreCalculator.cs
@@ -20,7 +20,7 @@ namespace Unity1Week202012
             }
         }
         
-        public int Evaluate(IEnumerable<PieceData> pieces)
+        public double Evaluate(IEnumerable<PieceData> pieces)
         {
             Dictionary<string, int> achievements = new Dictionary<string, int>();
             foreach(var combo in Services.Combinations)
